+++ src/main/java/sfa/product_service/dto/response/ProductPriceRes.java
@@ -10,8 +10,8 @@ import lombok.experimental.FieldDefaults;
 @AllArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class ProductPriceRes {
-    float warehousePrice;
-    float stockListPrice;
-    float retailerPrice;
-    float gstPercentage;
+    Double warehousePrice;
+    Double stockListPrice;
+    Double retailerPrice;
+    Double gstPercentage;
 }
